refactor: extract archive mixin from backup/service.py and provider schemas from config/schema.py
